Making Logging a separate module outside of autoconfiguration
Fixes: #1448.
