+++ spring-cloud-gcp-dependencies/pom.xml
@@ -83,6 +83,11 @@
 				<artifactId>spring-cloud-gcp-vision</artifactId>
 				<version>${project.version}</version>
 			</dependency>
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-gcp-logging</artifactId>
+				<version>${project.version}</version>
+			</dependency>
 
 			<!--Starters-->
 			<dependency>
