+++ spring-cloud-gcp-logging/src/main/java/org/springframework/cloud/gcp/logging/extractors/XCloudTraceIdExtractor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.gcp.autoconfigure.logging.extractors;
+package org.springframework.cloud.gcp.logging.extractors;
 
 import javax.servlet.http.HttpServletRequest;
