+++ spring-cloud-gcp-logging/src/main/java/org/springframework/cloud/gcp/logging/extractors/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Auto-configuration for Spring Cloud GCP Logging module's request trace ID extractors.
  */
-package org.springframework.cloud.gcp.autoconfigure.logging.extractors;
+package org.springframework.cloud.gcp.logging.extractors;
