+++ spring-cloud-gcp-logging/src/main/resources/org/springframework/cloud/gcp/logging/logback-appender.xml
@@ -10,7 +10,7 @@ Stackdriver appender logback configuration provided for import.
 
   <appender name="STACKDRIVER" class="com.google.cloud.logging.logback.LoggingAppender">
     <log>${STACKDRIVER_LOG_NAME}</log> <!-- Optional : default java.log -->
-    <enhancer>org.springframework.cloud.gcp.autoconfigure.logging.TraceIdLoggingEnhancer</enhancer>
+    <enhancer>org.springframework.cloud.gcp.logging.TraceIdLoggingEnhancer</enhancer>
     <flushLevel>${STACKDRIVER_LOG_FLUSH_LEVEL}</flushLevel> <!-- Optional : default ERROR -->
   </appender>
 </included>
