+++ spring-cloud-gcp-logging/src/test/resources/logback-test.xml
@@ -23,12 +23,4 @@
 		<appender-ref ref="STDOUT" />
 	</logger>
 
-	<if condition='property("it.logging").equalsIgnoreCase("true")'>
-		<then>
-			<include resource="org/springframework/cloud/gcp/logging/logback-appender.xml"/>
-			<logger name="org.springframework.cloud.gcp.autoconfigure.logging.it.StackdriverLoggingIntegrationTests" level="warn">
-				<appender-ref ref="STACKDRIVER" />
-			</logger>
-		</then>
-	</if>
 </configuration>
